Backend-add datetime

+++ src/main/java/com/example/demo/Controller/GrowthRecordController.java
@@ -4,6 +4,7 @@ import com.example.demo.DTO.GrowthRecord;
 import com.example.demo.DTO.KoiStatisticId;
 import com.example.demo.REQUEST_AND_RESPONSE.ReqResGrowth;
 import com.example.demo.Service.GrowthRecordService;
+import com.example.demo.Service.JWTUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -15,31 +16,43 @@ import java.time.LocalDate;
 public class GrowthRecordController {
     @Autowired
     private GrowthRecordService growService;
+    @Autowired
+    private JWTUtils jwt;
+
 
     @PostMapping("/{koiId}/addRecord")
-    ResponseEntity<ReqResGrowth> addRecord (@PathVariable Integer koiId, @RequestBody ReqResGrowth request){
-        return ResponseEntity.ok(growService.addRecord(request, koiId));
+    ResponseEntity<ReqResGrowth> addRecord (@RequestHeader ("Authorization") String token, @PathVariable Integer koiId, @RequestBody ReqResGrowth request){
+        int userId = jwt.extractUserId(token.replace("Bearer ", ""));
+        return ResponseEntity.ok(growService.addRecord(request, koiId, userId));
     }
 
     @GetMapping("/{koiId}/records")
-    ResponseEntity<ReqResGrowth> getRecords(@PathVariable Integer koiId){
-        return ResponseEntity.ok(growService.getrecords(koiId));
+    ResponseEntity<ReqResGrowth> getRecords(@RequestHeader ("Authorization") String token, @PathVariable Integer koiId){
+        int userId = jwt.extractUserId(token.replace("Bearer ", ""));
+        return ResponseEntity.ok(growService.getRecords(koiId, userId));
     }
     @GetMapping("/{koiId}/record/{date}")
-    ResponseEntity<ReqResGrowth> getRecord(@PathVariable Integer koiId, @PathVariable LocalDate date){
-        return  ResponseEntity.ok(growService.getRecord(koiId, date));
+    ResponseEntity<ReqResGrowth> getRecord(@RequestHeader ("Authorization") String token, @PathVariable Integer koiId, @PathVariable LocalDate date){
+        int userId = jwt.extractUserId(token.replace("Bearer ", ""));
+        return  ResponseEntity.ok(growService.getRecord(koiId, date, userId));
     }
 
-    @DeleteMapping("/{userId}/{koiId}/record/delete/{date}")
-    String deleteGrowth(@PathVariable Integer koiId, @PathVariable LocalDate date){
+    @DeleteMapping("/{koiId}/record/delete/{date}")
+    String deleteGrowth(@RequestHeader ("Authorization") String token, @PathVariable Integer koiId, @PathVariable LocalDate date){
+        int userId = jwt.extractUserId(token.replace("Bearer ", ""));
         KoiStatisticId id = new KoiStatisticId(date, koiId);
+        ReqResGrowth res = growService.getRecord(koiId, date, userId);
+        if(res.getStatusCode() ==200) {
             growService.deleteRecord(id);
             return "Delete success";
-
+        }else{
+            return "delete failed, " + res.getStatusCode() +": " + res.getError();
+        }
     }
 
-    @PutMapping("/{koiId}/record/update/{oldDate}")
-    ResponseEntity<ReqResGrowth> updateRecord(@PathVariable Integer koiId, @PathVariable LocalDate oldDate, @RequestBody ReqResGrowth request){
-        return ResponseEntity.ok(growService.updateRecord(request, koiId, oldDate));
+    @PutMapping("/{koiId}/record/update/{date}")
+    ResponseEntity<ReqResGrowth> updateRecord(@RequestHeader ("Authorization") String token, @PathVariable Integer koiId, @PathVariable LocalDate date, @RequestBody ReqResGrowth request){
+        int userId = jwt.extractUserId(token.replace("Bearer ", ""));
+        return ResponseEntity.ok(growService.updateRecord(request, koiId, date, userId));
     }
 }
